refactor: rename `trace()` to `executionTrace()` and create common models and internal utils
