+++ src/common/models/timer.model.ts
