+++ src/common/utils/safeError.spec.ts
