+++ src/common/utils/safeError.ts
