+++ src/index.ts
@@ -2,15 +2,16 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './common/models/engineNodeTrace.model';
+export * from './common/models/engineTrace.model';
+export * from './common/models/engineTraceOptions.model';
+export * from './common/models/executionTrace.model';
+export * from './common/models/timer.model';
 export * from './engine/executionEngine';
 export * from './engine/executionEngineDecorators';
 export * from './engine/traceableEngine';
-export * from './execute/execute';
+export * from './execution/execute';
 export * from './timer/executionTimer';
-export * from './timer/models/timer.model';
 export * from './trace/trace';
+export * from './trace/traceDecorator';
 export * from './trace/traceableExecution';
-export * from './trace/models/engineNodeTrace.model';
-export * from './trace/models/engineTrace.model';
-export * from './trace/models/engineTraceOptions.model';
-export * from './trace/models/executionTrace.model';
